Delete `CXCallable` and add static methods to `CXCallControllerManager`
